% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadiscretisation.R
\name{discretisation}
\alias{discretisation}
\title{Title For Discretisation}
\usage{
discretisation(df, var, method = c("equal_width", "equal_frequency"), bins = 4)
}
\arguments{
\item{bins}{}
}
\description{
Title For Discretisation
}
