% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Title Outlier Detector}
\usage{
find_outliers(df)
}
\arguments{
\item{df}{}
}
\value{
Prints out all outliers in the Data Set
}
\description{
Title Outlier Detector
}
