% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datanormalisation.R
\name{normalise}
\alias{normalise}
\title{Title Normalisation of Variable between 0 and 1}
\usage{
normalise(df, variable, new_min_A, new_max_A)
}
\arguments{
\item{new_max_A}{}
}
\description{
Title Normalisation of Variable between 0 and 1
}
